made progress with the chat app but I am having some issues updating the store for soime reason.
