+++ frontend/src/lib/Messenger.svelte
@@ -1,42 +1,32 @@
 <script lang="ts">
 //stores
-
 import ChatMessage from "./ChatMessage.svelte"
+import { user } from "../stores/user";
+import { sha256 } from "../utils/sha256";
+import { localMessages } from "../stores/messages";
 
-let localArray:string[]=[''] // WARN:very temp
 let chatMsg:string
 export function newMsg() {
-  console.log(chatMsg, localArray)
   if(chatMsg){
-    localArray.push(chatMsg)
-    chatMsg=''
-    }
+    const payload:Object = {
+      who:sha256($user.email),
+      data: chatMsg,
+      when: new Date()
+      }
 
+    localMessages.set(n => n.push(payload))
+    chatMsg=''
+    console.log(payload)
+    console.log($localMessages)
+
+    }
   }
 const onKeyPress = e => {
   if (e.charCode === 13) {
     newMsg()
   };
 };
-let user:object = {
-    who:"general-Kenobi",
-    when:"2020-01-20",
-  }
-let userMsg1:object = {
-    who:"general-Kenobi",
-    data:"Hello there",
-    when: new Date() 
 
-  }
-let user2:object = {
-    who:"laber-saber-robot-collector",
-  }
-let user2Msg1:object = {
-    who:"laber-saber-robot-collector",
-    data:"general-Kenobi!",
-    when: new Date()
-
-  }
 </script>
 
 <template>
@@ -49,13 +39,20 @@ let user2Msg1:object = {
       on:keypress={onKeyPress}
       type="text" name="msg-chat" id="msg-chat" class="shadow-sm focus:ring-indigo-500 focus:border-indigo-500 block w-full sm:text-sm border-gray-300 rounded-md" placeholder="...">
     </div>
+    <img
+      class="h-8 w-8 rounded-full"
+      src="{`http://tinygraphs.com/labs/isogrids/hexa16/${sha256(
+        $user.email
+      )}?theme=seascape&numcolors=4`}"
+      alt={$user.email} />
   </div>
-  <ul>
   <!--TODO: check that the chat is showing -->
+  {#each $localMessages as message}
+    <svelte:component this={ChatMessage} message={message.message} sender={message.sender} />
+  {/each}
+  <!-- 
   <ChatMessage message={userMsg1} sender={user2.who} />
   <ChatMessage message={user2Msg1} sender={user2.who} />
-  {#each localArray as entry}
-    <li>{entry}</li>
-  {/each}
-  </ul>
+  -->
+
 </template>
